services/agent_service/requirements.txt
google-adk
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
python-dotenv
